+++ x/perp/v2/keeper/liquidate_test.go
@@ -160,6 +160,12 @@ func TestMultiLiquidate(t *testing.T) {
 						},
 					),
 				),
+				ContainsLiquidateEvent(types.LiquidationFailedEvent{
+					Pair:       pairBtcUsdc,
+					Trader:     alice.String(),
+					Liquidator: liquidator.String(),
+					Reason:     types.LiquidationFailedEvent_POSITION_HEALTHY,
+				}),
 			),
 
 		TC("mixed bag").
@@ -215,6 +221,25 @@ func TestMultiLiquidate(t *testing.T) {
 						},
 					),
 				),
+
+				ContainsLiquidateEvent(types.LiquidationFailedEvent{
+					Pair:       pairAtomUsdc,
+					Trader:     alice.String(),
+					Liquidator: liquidator.String(),
+					Reason:     types.LiquidationFailedEvent_POSITION_HEALTHY,
+				}),
+				ContainsLiquidateEvent(types.LiquidationFailedEvent{
+					Pair:       pairSolUsdc,
+					Trader:     alice.String(),
+					Liquidator: liquidator.String(),
+					Reason:     types.LiquidationFailedEvent_NONEXISTENT_PAIR,
+				}),
+				ContainsLiquidateEvent(types.LiquidationFailedEvent{
+					Pair:       pairBtcUsdc,
+					Trader:     bob.String(),
+					Liquidator: liquidator.String(),
+					Reason:     types.LiquidationFailedEvent_NONEXISTENT_POSITION,
+				}),
 			),
 	}